refactor: separate cli from index.js
